// src/lib/api/model/usersControllerSearchUsersClass.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Wildcat Radio v2 API
 * Control plane — see docs/final-build-plan/03-API-CONTRACT.md
 * OpenAPI spec version: 0.0.1
 */

export type UsersControllerSearchUsersClass = typeof UsersControllerSearchUsersClass[keyof typeof UsersControllerSearchUsersClass];


export const UsersControllerSearchUsersClass = {
  CAMPUS: 'CAMPUS',
  GUEST: 'GUEST',
} as const;